Add delete wallet button
