+++ lib/src/domain/usecases/delete_wallet_usecase.dart
@@ -1,15 +1,13 @@
 import 'package:donationwallet/src/data/repositories/wallet_repository.dart';
-import 'package:donationwallet/src/domain/entities/wallet_entity.dart';
 
 class DeleteWalletUseCase {
   final WalletRepository walletRepository;
 
   DeleteWalletUseCase(this.walletRepository);
 
-  Future<WalletEntity?> call(String label) async {
+  Future<bool> call(String label) async {
     try {
-      final res = await walletRepository.getWallet(label);
-      return res;
+      return await walletRepository.rmWallet(label);
     } catch (e) {
       rethrow;
     }
